Moves components to the proper directory
