+++ raiden-dapp/src/components/Divider.vue
