+++ raiden-dapp/src/components/TokenInformation.vue
